Crate dcso3

Modules

airbase

attribute

coalition

controller

coord

country

dcs

env

event

group

hooks

land

lfs

mission_commands

net

object

perf

spot

static_object

timer

trigger

unit

warehouse

weapon

world

Macros

atomic_id

bitflags_enum

path

record_perf

simple_enum

string_enum

wrapped_prim

wrapped_table

Structs

Box3

Color

HooksLua

LuaVec2

LuaVec3

MizLua

Path

Position3

Quad2

Sequence

String

Time

Enums

PathElt

VolumeType

Traits

DcsTableExt

DeepClone

LuaEnv

Functions

as_tbl

as_tbl_ref

azumith2d

azumith2d_to

azumith3d

azumith3d_to

centroid2d

centroid3d

change_heading

change the heading (in radians) by the specified amount, adjusting if it rotates through 0/2 pi. e.g. change_heading(3/2 pi, pi) -> pi / 2 not 5 / 2 pi

create_root_module

cvt_err

degrees_to_radians

err

is_hooks_env

lua_err

normal2

pointing_towards2

return a unit vector pointing in the specified direction. angle is in radians

radians_to_degrees

rotate2d

Rotate a collection of points in 2d space around their center point keeping the relative orientations of the points constant. The angle is in radians.

rotate2d_gen

Rotate a collection of points in 2d space around their center point keeping the relative orientations of the points constant. The angle is in radians. General about the underlying point container type

rotate2d_vec

Same as rotate2d, but construct and return a vec containing the rotated points in the same order as the they appear in the input slice.

value_to_json

wrap

wrap_f

Type Aliases

Vector2

Vector3